Rename drawing/Weather.h -> drawing/WeatherDrawer.h
