+++ src/openrct2-ui/drawing/engines/opengl/OpenGLDrawingEngine.cpp
@@ -35,7 +35,7 @@
     #include <openrct2/drawing/IDrawingEngine.h>
     #include <openrct2/drawing/InvalidationGrid.h>
     #include <openrct2/drawing/LightFX.h>
-    #include <openrct2/drawing/Weather.h>
+    #include <openrct2/drawing/WeatherDrawer.h>
     #include <openrct2/interface/Screenshot.h>
     #include <openrct2/ui/UiContext.h>
     #include <openrct2/world/Weather.h>
