+++ src/openrct2/drawing/WeatherDrawer.cpp
@@ -7,7 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
-#include "Weather.h"
+#include "WeatherDrawer.h"
 
 #include "../Game.h"
 #include "../GameState.h"
