+++ src/openrct2/drawing/WeatherDrawer.h
