+++ src/openrct2/libopenrct2.vcxproj
@@ -268,10 +268,10 @@
     <ClInclude Include="drawing\Rectangle.h" />
     <ClInclude Include="drawing\RenderTarget.h" />
     <ClInclude Include="drawing\ScrollingText.h" />
-    <ClInclude Include="drawing\Weather.h" />
     <ClInclude Include="drawing\Text.h" />
     <ClInclude Include="drawing\TextColour.h" />
     <ClInclude Include="drawing\TTF.h" />
+    <ClInclude Include="drawing\WeatherDrawer.h" />
     <ClInclude Include="drawing\X8DrawingEngine.h" />
     <ClInclude Include="Editor.h" />
     <ClInclude Include="EditorObjectSelectionSession.h" />
@@ -632,7 +632,6 @@
     <ClInclude Include="scripting\bindings\ride\ScTrackIterator.h" />
     <ClInclude Include="scripting\bindings\ride\ScTrackSegment.h" />
     <ClInclude Include="scripting\bindings\world\ScAward.hpp" />
-    <ClInclude Include="scripting\bindings\world\ScWeather.hpp" />
     <ClInclude Include="scripting\bindings\world\ScDate.hpp" />
     <ClInclude Include="scripting\bindings\world\ScMap.hpp" />
     <ClInclude Include="scripting\bindings\world\ScPark.hpp" />
@@ -641,6 +640,7 @@
     <ClInclude Include="scripting\bindings\world\ScScenario.hpp" />
     <ClInclude Include="scripting\bindings\world\ScTile.hpp" />
     <ClInclude Include="scripting\bindings\world\ScTileElement.hpp" />
+    <ClInclude Include="scripting\bindings\world\ScWeather.hpp" />
     <ClInclude Include="scripting\Duktape.hpp" />
     <ClInclude Include="scripting\HookEngine.h" />
     <ClInclude Include="scripting\IconNames.hpp" />
@@ -657,7 +657,6 @@
     <ClInclude Include="windows\Intent.h" />
     <ClInclude Include="windows\TileInspectorGlobals.h" />
     <ClInclude Include="world\Banner.h" />
-    <ClInclude Include="world\Weather.h" />
     <ClInclude Include="world\ConstructionClearance.h" />
     <ClInclude Include="world\Entrance.h" />
     <ClInclude Include="world\Footpath.h" />
@@ -676,6 +675,7 @@
     <ClInclude Include="world\TileElementsView.h" />
     <ClInclude Include="world\TileInspector.h" />
     <ClInclude Include="world\TilePointerIndex.hpp" />
+    <ClInclude Include="world\Weather.h" />
     <ClInclude Include="world\map_generator\HeightMap.hpp" />
     <ClInclude Include="world\map_generator\MapGen.h" />
     <ClInclude Include="world\map_generator\MapHelpers.h" />
@@ -866,7 +866,6 @@
     <ClCompile Include="drawing\LightFX.cpp" />
     <ClCompile Include="drawing\Line.cpp" />
     <ClCompile Include="drawing\NewDrawing.cpp" />
-    <ClCompile Include="drawing\Weather.cpp" />
     <ClCompile Include="drawing\Rectangle.cpp" />
     <ClCompile Include="drawing\RenderTarget.cpp" />
     <ClCompile Include="drawing\ScrollingText.cpp" />
@@ -875,6 +874,7 @@
     <ClCompile Include="drawing\TextColour.cpp" />
     <ClCompile Include="drawing\TTF.cpp" />
     <ClCompile Include="drawing\TTFSDLPort.cpp" />
+    <ClCompile Include="drawing\WeatherDrawer.cpp" />
     <ClCompile Include="drawing\X8DrawingEngine.cpp" />
     <ClCompile Include="Editor.cpp" />
     <ClCompile Include="EditorObjectSelectionSession.cpp" />
